+++ demo/src/screens/componentScreens/index.js
@@ -31,6 +31,7 @@ import WheelPickerDialogScreen from './WheelPickerDialogScreen';
 import SliderScreen from './SliderScreen';
 import FloatingButtonScreen from './FloatingButtonScreen';
 import ColorPickerScreen from './ColorPickerScreen';
+import StackAggregatorScreen from './StackAggregatorScreen';
 
 
 Navigation.registerComponent('unicorn.components.ActionSheetScreen', () => ActionSheetScreen);
@@ -65,3 +66,4 @@ Navigation.registerComponent('unicorn.components.WheelPickerDialogScreen', () =>
 Navigation.registerComponent('unicorn.components.SliderScreen', () => SliderScreen);
 Navigation.registerComponent('unicorn.components.FloatingButtonScreen', () => FloatingButtonScreen);
 Navigation.registerComponent('unicorn.components.ColorPickerScreen', () => ColorPickerScreen);
+Navigation.registerComponent('unicorn.components.StackAggregatorScreen', () => StackAggregatorScreen);
